pyproject.toml
[tool.poetry]
name = "my-django-project"
version = "0.1.0"
description = ""
authors = ["XUeshuo Li <spoonlee24k@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
django-ninja = "^1.3.0"
psycopg2-binary = "^2.9.3"
kedro = "^0.19.9"
pandas = "^2.2.3"
kedro-datasets = "^5.1.0"
redis = "^5.2.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
django-environ = "^0.11.2"
boto3 = "^1.35.58"
faker = "^33.0.0"
openpyxl = "^3.1.5"
django-cors-headers = "^4.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
